+++ sdk/src/protos_json/led_proto.hh
@@ -22,8 +22,8 @@ struct VoiceData {
 };
 
 struct FanData {
-  int fan{};
-  NLOHMANN_DEFINE_TYPE_INTRUSIVE(FanData, fan)
+  int mode{};
+  NLOHMANN_DEFINE_TYPE_INTRUSIVE(FanData, mode)
 };
 
 struct LedStyleLedData {